Union Router and Panel to Router
