+++ router/mod.ts
@@ -1,3 +1,2 @@
 export { MemoryHistory } from './memory_history.ts'
-export { Router } from './router.ts'
-export { Panel } from './panel.ts'
+export { walk } from './walk.ts'
